Extract auth into component; Setup landing page routing;
