+++ src/routes/dashboard.jsx
@@ -1,7 +1,6 @@
 export default function Dashboard() {
   return (
-    <main style={{ padding: "1rem 0" }}>
-      <h2>Invoices</h2>
+    <main>
     </main>
   );
 }
